Renamed MockTree to MockDomain for consistancy.
