+++ src/components/TreeView/tests/TreeViewComponent.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import renderer, { ReactTestRenderer } from "react-test-renderer";
 import TreeViewComponent, { TreeView } from "../TreeViewComponent";
 import SemanticDomainWithSubdomains from "../SemanticDomain";
-import mockTree from "./MockSemanticTree";
+import MockDomain from "./MockSemanticDomain";
 import { store } from "../../../store";
 import { Provider } from "react-redux";
 
@@ -30,7 +30,7 @@ jest.mock("../TreeViewReducer", () => {
   return {
     ...realReducer,
     createDomains: () => {
-      return { currentDomain: mockTree };
+      return { currentDomain: MockDomain };
     },
   };
 });
@@ -51,11 +51,11 @@ describe("Tests AddWords", () => {
     snapTest("default view");
   });
 
-  // This assumes mockTree was used in createTree() below and mockTree.name !== ""
+  // This assumes MockDomain was used in createTree() below and MockDomain.name !== ""
   it("Navigates to .currentDomain in construction", () => {
     createTree();
     expect(NAVIGATE_MOCK).toHaveBeenCalledTimes(1);
-    expect(NAVIGATE_MOCK).toHaveBeenCalledWith(mockTree);
+    expect(NAVIGATE_MOCK).toHaveBeenCalledWith(MockDomain);
   });
 
   it("Sets a new domain upon navigation", () => {
@@ -73,7 +73,7 @@ describe("Tests AddWords", () => {
   });
 
   it("Returns control to caller when the same semantic domain is passed in", () => {
-    treeHandle.animate(mockTree);
+    treeHandle.animate(MockDomain);
     jest.runAllTimers();
 
     expect(RETURN_MOCK).toHaveBeenCalledTimes(1);
@@ -85,7 +85,7 @@ function createTree() {
     treeMaster = renderer.create(
       <Provider store={store}>
         <TreeViewComponent
-          currentDomain={mockTree}
+          currentDomain={MockDomain}
           returnControlToCaller={RETURN_MOCK}
           navigateTree={NAVIGATE_MOCK}
         />
