+++ src/components/TreeView/tests/TreeViewHeader.test.tsx
@@ -3,7 +3,7 @@ import renderer, { ReactTestRenderer } from "react-test-renderer";
 
 import SemanticDomainWithSubdomains from "../SemanticDomain";
 import TreeViewHeader from "../TreeViewHeader";
-import MockTree from "./MockSemanticTree";
+import MockDomain from "./MockSemanticDomain";
 
 // Variable event
 var event = {
@@ -33,7 +33,7 @@ window = {
 };
 
 beforeEach(() => {
-  setTree(MockTree.subdomains[1]);
+  setTree(MockDomain.subdomains[1]);
   MOCK_ANIMATE.mockClear();
 });
 
@@ -54,11 +54,11 @@ describe("Tests TreeViewHeader", () => {
 
   // onKeyDown
   it("Search & select domain switches semantic domain if given number found", () => {
-    treeHandle.setState({ input: MockTree.id });
+    treeHandle.setState({ input: MockDomain.id });
     event.target.value = "not empty";
     treeHandle.searchAndSelectDomain((event as any) as React.KeyboardEvent);
 
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockTree);
+    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain);
     expect(treeHandle.state.input).toEqual("");
     expect(event.target.value).toEqual("");
   });
@@ -75,11 +75,11 @@ describe("Tests TreeViewHeader", () => {
   });
 
   it("Search & select domain switches semantic domain if given name found", () => {
-    treeHandle.setState({ input: MockTree.subdomains[2].name });
+    treeHandle.setState({ input: MockDomain.subdomains[2].name });
     event.target.value = "not empty";
     treeHandle.searchAndSelectDomain((event as any) as React.KeyboardEvent);
 
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockTree.subdomains[2]);
+    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[2]);
     expect(treeHandle.state.input).toEqual("");
     expect(event.target.value).toEqual("");
   });
@@ -98,15 +98,15 @@ describe("Tests TreeViewHeader", () => {
   // getBrotherDomain
   it("provides the proper brother domains", () => {
     // Standard navigation
-    expect(treeHandle.getBrotherDomain(-1)).toEqual(MockTree.subdomains[0]);
-    expect(treeHandle.getBrotherDomain(1)).toEqual(MockTree.subdomains[2]);
+    expect(treeHandle.getBrotherDomain(-1)).toEqual(MockDomain.subdomains[0]);
+    expect(treeHandle.getBrotherDomain(1)).toEqual(MockDomain.subdomains[2]);
 
     // Check with indices out-of-bounds
     expect(treeHandle.getBrotherDomain(-2)).toEqual(undefined);
     expect(treeHandle.getBrotherDomain(2)).toEqual(undefined);
 
     // Check that a domain w/ no parentDomain has no brotherDomains
-    setTree(MockTree);
+    setTree(MockDomain);
     expect(treeHandle.getBrotherDomain(-1)).toEqual(undefined);
   });
 
@@ -114,19 +114,19 @@ describe("Tests TreeViewHeader", () => {
   it("navigateKeys w/ Arrow Down switches to parent domain", () => {
     event.key = "ArrowDown";
     treeHandle.navigateDomainArrowKeys((event as any) as KeyboardEvent);
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockTree);
+    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain);
   });
 
   it("navigateKeys w/ Arrow Left switches to left brother domain", () => {
     event.key = "ArrowLeft";
     treeHandle.navigateDomainArrowKeys((event as any) as KeyboardEvent);
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockTree.subdomains[0]);
+    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[0]);
   });
 
   it("navigateKeys w/ Arrow Right switches to right brother domain", () => {
     event.key = "ArrowRight";
     treeHandle.navigateDomainArrowKeys((event as any) as KeyboardEvent);
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockTree.subdomains[2]);
+    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[2]);
   });
 });
